add compatibility bar tests
